# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/extract-cells/{job_id}": {
      "get": {
        "tags": [
          "cellextraction"
        ],
        "summary": "Get Extract Cells Status",
        "operationId": "get_extract_cells_status_api_v1_extract_cells__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/extract-cells": {
      "post": {
        "tags": [
          "cellextraction"
        ],
        "summary": "Extract Cells",
        "operationId": "extract_cells_api_v1_extract_cells_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractCellsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExtractCellsRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "layer_mode": {
            "type": "string",
            "title": "Layer Mode"
          },
          "param1": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Param1",
            "default": 130
          },
          "image_size": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Image Size",
            "default": 200
          }
        },
        "type": "object",
        "required": [
          "filename",
          "layer_mode"
        ],
        "title": "ExtractCellsRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}